SFB-121: create util method to check if the displayType can have form:options
